+++ lib/manageiq/loggers/json_logger.rb
@@ -63,13 +63,7 @@ module ManageIQ
         end
 
         def request_id
-          Thread.current[:request_id] || Thread.current[:current_request]&.request_id.tap do |request_id|
-            if request_id
-              require "active_support"
-              require "active_support/deprecation"
-              ActiveSupport::Deprecation.warn("Usage of `Thread.current[:current_request]&.request_id` will be deprecated in version 0.5.0. Please switch to `Thread.current[:request_id]` to log request_id automatically.")
-            end
-          end
+          Thread.current[:request_id]
         end
       end
     end